- migrate and merge railties and tasks - add warning to `require “delayed_job”` and `require “delayed_job_active_record”`
