+++ lib/delayed_job.rb
@@ -1,3 +1,6 @@
+$stderr.puts "require 'delayed_job' is deprecated, migrate to require 'tom_queue'"
+caller.each { |e| $stderr.puts "\t#{e}"}
+
 require 'active_support'
 require 'delayed/exceptions'
 require 'delayed/message_sending'
